Modify the RemoteComm connection handshake
Changed the way in which the handshake is performed between
the RemoteComm class the control software, not button presses
are necessary, also the main loop now checks for disconnections
and will attempt to reconnect when this is detected.

Added some documentation to functions in main_mob_platform.ino and
removed unused code.